move store request form into a component in its own file. parent page that renders the form is a server component
